+++ src/GregClient/GregClient.cs
@@ -10,9 +10,7 @@ namespace Greg
     public class GregClient : IGregClient
     {
         private readonly RestClient _client;
-
         public string BaseUrl { get { return _client.Options.BaseUrl.ToString(); } }
-
         public readonly IAuthProvider _authProvider;
         public IAuthProvider AuthProvider
         {