+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * switch from newtonsoft.json to system.text.json
 * add unit test project
 * drop all targets below net6.
+* OAuth1 is no longer supported, we no longer sign requests.
+* Request base classes are modified to remove support for signing requests.
 ### version 2.5
 * added user/votes request route
 * Add net6.0 as a target framework