+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### version 3.0
+* update RestSharp to 109.0.1, this comes with many API breaking changes, some affecting us, see below:
+* removal of basic auth - our API no longer makes sense with RestSharp immutable clients.
+* switch from newtonsoft.json to system.text.json
+* add unit test project
+* drop all targets below net6.
+* OAuth1 is no longer supported, we no longer sign requests.
+* Request base classes are modified to remove support for signing requests.
 ### version 2.5
 * added user/votes request route
 * Add net6.0 as a target framework